Refactored the whole app.rb to avoid having a 500 long file
