+++ .github/workflows/continuous_delivery.yml
@@ -2,7 +2,7 @@ name: Continuous Delivery
 
 on:
   push:
-    branches: [ "main", "feature/docker-update" ]  # Trigger on pushes to main branch
+    branches: [ "main", "feature/refactor" ]  # Trigger on pushes to main branch
 
 jobs:
   build-and-push:
